+++ src/ragtime/config.py
@@ -39,6 +39,7 @@ DEFAULT_HUMAN_EVAL_COL:int = DEFAULT_HUMAN_EVAL_COL if "DEFAULT_HUMAN_EVAL_COL" 
 # # # LLMs
 DEFAULT_LITELLM_RETRIES:int = DEFAULT_LITELLM_RETRIES if "DEFAULT_LITELLM_RETRIES" in globals() else None
 DEFAULT_LITELLM_TEMP:int = DEFAULT_LITELLM_TEMP if "DEFAULT_LITELLM_TEMP" in globals() else None
+DEFAULT_MAX_TOKENS:int = DEFAULT_MAX_TOKENS if "DEFAULT_MAX_TOKENS" in globals() else None
 
 # Logging - class to add msg
 class RagtimeLogger(logging.LoggerAdapter):
@@ -59,7 +60,7 @@ def init(root_folder:Path):
     global DEFAULT_SPREADSHEET_TEMPLATE, DEFAULT_WORKSHEET, DEFAULT_HEADER_SIZE, DEFAULT_QUESTION_COL, DEFAULT_FACTS_COL
     global DEFAULT_ANSWERS_COL, DEFAULT_HUMAN_EVAL_COL
     # # # LLMs
-    global DEFAULT_LITELLM_RETRIES, DEFAULT_LITELLM_TEMP
+    global DEFAULT_LITELLM_RETRIES, DEFAULT_LITELLM_TEMP, DEFAULT_MAX_TOKENS
     # Logger
     global logger
     
@@ -87,6 +88,7 @@ def init(root_folder:Path):
     # # # LLMs
     DEFAULT_LITELLM_RETRIES = 3
     DEFAULT_LITELLM_TEMP = 0
+    DEFAULT_MAX_TOKENS = 1000 # empirically noticed the biggest answers are 4000 characters long - and 1 token is between 4 and 5 chars - keep the largest value, i.e. 4 chars per token
 
     ####################
     # LOGGING